lambda/requirements.txt

bedrock-agentcore-starter-toolkit>=0.1.10
bedrock-agentcore>=0.1.3

mcp>=1.13.1

# Strands Agents libraries
strands-agents>=1.8.0
strands-agents-tools>=0.2.7

# MCP Server libraries
awslabs.aws-serverless-mcp-server
awslabs-aws-api-mcp-server>=0.2.13
awslabs-aws-documentation-mcp-server>=0.1.7

aws-cdk-lib>=2.214.0
boto3>=1.34.131